Revert "orphan window"
This reverts commit 3ebcd589a47a13c76d0fa471323a4f77059e5187.
