+++ DockDoor/consts.swift
@@ -94,5 +94,201 @@ extension Defaults.Keys {
     static let appNameFilters = Key<[String]>("appNameFilters", default: [])
     static let windowTitleFilters = Key<[String]>("windowTitleFilters", default: [])
     static let customAppDirectories = Key<[String]>("customAppDirectories", default: [])
-    static let orphanedWindowAssociations = Key<[OrphanedWindowAssociation]>("orphanedWindowAssociations", default: [])
+}
+
+// MARK: Display Configurations
+
+enum WindowTitleDisplayCondition: String, CaseIterable, Defaults.Serializable {
+    case all
+    case dockPreviewsOnly
+    case windowSwitcherOnly
+
+    var localizedName: String {
+        switch self {
+        case .all:
+            String(localized: "Dock Previews & Window Switcher", comment: "Preview window title display condition option")
+        case .dockPreviewsOnly:
+            String(localized: "Dock Previews only", comment: "Preview window title condition display option")
+        case .windowSwitcherOnly:
+            String(localized: "Window Switcher only", comment: "Preview window title condition display option")
+        }
+    }
+}
+
+enum WindowTitlePosition: String, CaseIterable, Defaults.Serializable {
+    case bottomLeft
+    case bottomRight
+    case topRight
+    case topLeft
+
+    var localizedName: String {
+        switch self {
+        case .bottomLeft:
+            String(localized: "Bottom Left", comment: "Preview window title position option")
+        case .bottomRight:
+            String(localized: "Bottom Right", comment: "Preview window title position option")
+        case .topRight:
+            String(localized: "Top Right", comment: "Preview window title position option")
+        case .topLeft:
+            String(localized: "Top Left", comment: "Preview window title position option")
+        }
+    }
+}
+
+enum AppNameStyle: String, CaseIterable, Defaults.Serializable {
+    case `default`
+    case shadowed
+    case popover
+
+    var localizedName: String {
+        switch self {
+        case .default:
+            String(localized: "Default", comment: "Preview title style option")
+        case .shadowed:
+            String(localized: "Shadowed", comment: "Preview title style option")
+        case .popover:
+            String(localized: "Popover", comment: "Preview title style option")
+        }
+    }
+}
+
+enum WindowTitleVisibility: String, CaseIterable, Defaults.Serializable {
+    case whenHoveringPreview
+    case alwaysVisible
+
+    var localizedName: String {
+        switch self {
+        case .whenHoveringPreview:
+            String(localized: "When hovering over the preview", comment: "Window title visibility option")
+        case .alwaysVisible:
+            String(localized: "Always visible", comment: "Window title visibility option")
+        }
+    }
+}
+
+enum TrafficLightButtonsVisibility: String, CaseIterable, Defaults.Serializable {
+    case never
+    case dimmedOnPreviewHover
+    case fullOpacityOnPreviewHover
+    case alwaysVisible
+
+    var localizedName: String {
+        switch self {
+        case .never:
+            String(localized: "Never visible", comment: "Traffic light buttons visibility option")
+        case .dimmedOnPreviewHover:
+            String(localized: "On window hover; Dimmed until button hover", comment: "Traffic light buttons visibility option")
+        case .fullOpacityOnPreviewHover:
+            String(localized: "On window hover; Full opacity", comment: "Traffic light buttons visibility option")
+        case .alwaysVisible:
+            String(localized: "Always visible; Full opacity", comment: "Traffic light buttons visibility option")
+        }
+    }
+}
+
+enum TrafficLightButtonsPosition: String, CaseIterable, Defaults.Serializable {
+    case topLeft
+    case topRight
+    case bottomRight
+    case bottomLeft
+
+    var localizedName: String {
+        switch self {
+        case .topLeft:
+            String(localized: "Top Left", comment: "Traffic light buttons position option")
+        case .topRight:
+            String(localized: "Top Right", comment: "Traffic light buttons position option")
+        case .bottomRight:
+            String(localized: "Bottom Right", comment: "Traffic light buttons position option")
+        case .bottomLeft:
+            String(localized: "Bottom Left", comment: "Traffic light buttons position option")
+        }
+    }
+}
+
+enum WindowSwitcherPlacementStrategy: String, CaseIterable, Defaults.Serializable {
+    case screenWithMouse
+    case screenWithLastActiveWindow
+    case pinnedToScreen
+
+    var localizedName: String {
+        switch self {
+        case .screenWithMouse:
+            String(localized: "Screen with mouse", comment: "Window switcher placement option")
+        case .screenWithLastActiveWindow:
+            String(localized: "Screen with last active window", comment: "Window switcher placement option")
+        case .pinnedToScreen:
+            String(localized: "Pinned to screen", comment: "Window switcher placement option")
+        }
+    }
+}
+
+enum WindowSwitcherControlPosition: String, CaseIterable, Defaults.Serializable {
+    case topLeading
+    case topTrailing
+    case bottomLeading
+    case bottomTrailing
+
+    var localizedName: String {
+        switch self {
+        case .topLeading:
+            String(localized: "At top - Title on left, controls on right")
+        case .topTrailing:
+            String(localized: "At top - Controls on left, title on right")
+        case .bottomLeading:
+            String(localized: "At bottom - Title on left, controls on right")
+        case .bottomTrailing:
+            String(localized: "At bottom - Controls on left, title on right")
+        }
+    }
+}
+
+// MARK: Action Configurations
+
+enum PreviewHoverAction: String, CaseIterable, Defaults.Serializable {
+    case none
+    case tap
+    case previewFullSize
+
+    var localizedName: String {
+        switch self {
+        case .none:
+            String(localized: "No action", comment: "Window popup hover action option")
+        case .tap:
+            String(localized: "Simulate a click (open the window)", comment: "Window popup hover action option")
+        case .previewFullSize:
+            String(localized: "Present a full size preview of the window", comment: "Window popup hover action option")
+        }
+    }
+}
+
+enum AeroShakeAction: String, CaseIterable, Defaults.Serializable {
+    case none
+    case all
+    case except
+
+    var localizedName: String {
+        switch self {
+        case .none:
+            String(localized: "No action", comment: "Aero shake action option")
+        case .all:
+            String(localized: "Minimize all windows", comment: "Aero shake action option")
+        case .except:
+            String(localized: "Minimize all windows except the current one", comment: "Aero shake action option")
+        }
+    }
+}
+
+enum DockClickAction: String, CaseIterable, Defaults.Serializable {
+    case minimize
+    case hide
+
+    var localizedName: String {
+        switch self {
+        case .minimize:
+            String(localized: "Minimize windows", comment: "Dock click action option")
+        case .hide:
+            String(localized: "Hide application", comment: "Dock click action option")
+        }
+    }
 }
